+++ meta/3rd/love2d/config.lua
@@ -1,0 +1,8 @@
+words   = {'love%.%w+'}
+configs = {
+    {
+        key    = 'Lua.runtime.version',
+        action = 'set',
+        value  = 'LuaJIT',
+    },
+}